move scene files to own folder
