+++ src/scene/scene.js
@@ -1,30 +1,30 @@
-import log from './utils/log';
-import Utils from './utils/utils';
-import debugSettings from './utils/debug_settings';
-import * as URLs from './utils/urls';
-import WorkerBroker from './utils/worker_broker';
-import Task from './utils/task';
-import subscribeMixin from './utils/subscribe';
-import sliceObject from './utils/slice';
-import Context from './gl/context';
-import Texture from './gl/texture';
-import ShaderProgram from './gl/shader_program';
-import VertexArrayObject from './gl/vao';
-import {StyleManager} from './styles/style_manager';
-import {Style} from './styles/style';
-import StyleParser from './styles/style_parser';
+import log from '../utils/log';
+import Utils from '../utils/utils';
+import debugSettings from '../utils/debug_settings';
+import * as URLs from '../utils/urls';
+import WorkerBroker from '../utils/worker_broker';
+import Task from '../utils/task';
+import subscribeMixin from '../utils/subscribe';
+import sliceObject from '../utils/slice';
+import Context from '../gl/context';
+import Texture from '../gl/texture';
+import ShaderProgram from '../gl/shader_program';
+import VertexArrayObject from '../gl/vao';
+import {StyleManager} from '../styles/style_manager';
+import {Style} from '../styles/style';
+import StyleParser from '../styles/style_parser';
 import SceneLoader from './scene_loader';
-import View from './view';
-import Light from './light';
-import Tile from './tile/tile';
-import TileManager from './tile/tile_manager';
-import DataSource from './sources/data_source';
-import './sources/sources';
-import FeatureSelection from './selection';
-import RenderStateManager from './gl/render_state';
-import CanvasText from './styles/text/canvas_text';
-import FontManager from './styles/text/font_manager';
-import MediaCapture from './utils/media_capture';
+import View from '../view';
+import Light from '../light';
+import Tile from '../tile/tile';
+import TileManager from '../tile/tile_manager';
+import DataSource from '../sources/data_source';
+import '../sources/sources';
+import FeatureSelection from '../selection';
+import RenderStateManager from '../gl/render_state';
+import CanvasText from '../styles/text/canvas_text';
+import FontManager from '../styles/text/font_manager';
+import MediaCapture from '../utils/media_capture';
 
 // Load scene definition: pass an object directly, or a URL as string to load remotely
 export default class Scene {
